requirements.txt
# --- GUI Framework ---
# Указываем конкретные СТАБИЛЬНЫЕ версии
kivy==2.3.0
kivymd==1.2.0

# --- Networking & Scraping ---
requests>=2.31.0
beautifulsoup4>=4.12.0
duckduckgo-search==5.0.0
aiohttp>=3.9.0

# --- AI & Services ---
googletrans==4.0.0-rc1
gTTS>=2.5.0

google-generativeai>=0.4.0 # Новая библиотека для работы с Gemini